refactor: StatusView 생성
